+++ src/lib/components/main-page/game-page/QueueDisplay.svelte
