+++ src/lib/components/main-page/game-page/QueueItem.svelte
