+++ src/lib/components/main-page/game-page/Reset.svelte
