+++ src/lib/components/main-page/game-page/TeamManagement.svelte
