+++ src/lib/components/main-page/game-page/TimeSelector.svelte
