+++ src/lib/components/main-page/game-page/Timer.svelte
