+++ src/lib/components/main-page/table-page/LeagueTable.svelte
